Versión 4.1.1
- Corrección de Mepro. Sólo se podía editar, ejecutar y eliminar el
último script de la lista
- Nuevo script que muestra el reporte de faltantes de inventario,
también conocida como lista de resurtimiento.
- Corrección del cuadro de diálogo de mensajes de error: Los detalles
del error ahora pueden verse maximizados.
- Mejora al cuadro de diálogo de mensaje de error: Botón para copiar
los detalles (stacktrace) de un error al portapapeles
- Retoque de las plantillas de etiquetas
- Corrección del filtro de cortes de sucursal, no se podía filtrar por
fecha
- Corrección del error que no permitía realizar cortes de caja (lanzaba
un error)
- Corrección del panel de autenticación con NIP. Ahora se puede
utilizar la tecla <Enter>.

+++ main/src/omoikane/sistema/cortes/EstrategiaEstandar.groovy
@@ -150,11 +150,11 @@ public class EstrategiaEstandar extends EstrategiasCorte {
                 corte.setDesde( desde );
                 corte.setHasta( hasta );
                 corte.setIdCaja( IDCaja );
-                corte.setSucursalId( IDAlmacen );
+                corte.setSucursalId( IDAlmacen as int );
                 corte.setDepositos( depositos );
                 corte.setRetiros( retiros );
-                corte.setFolioInicial( folios.ini );
-                corte.setFolioFinal( folios.fin );
+                corte.setFolioInicial( folios.ini as Long);
+                corte.setFolioFinal( folios.fin as Long);
                 corte.setAbierto(false);
 
                 List<CorteImpuesto> corteImpuestoList = []
